[-] Correção para compilação em lazarus x64.
git-svn-id: svn://svn.code.sf.net/p/acbr/code/trunk2@14508 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Fontes/ACBrDFe/ACBrNFSe/PCNNFSe/pnfsNFSeR.pas
@@ -3142,8 +3142,8 @@ begin
           NFSe.Servico.ItemServico.Items[I - 1].Aliquota      := Leitor.rCampo(tcDe2, 'Aliquota');
           NFSe.Servico.ItemServico.Items[I - 1].ValorServicos := Leitor.rCampo(tcDe2, 'ValorServico');
           NFSe.Servico.ItemServico.Items[I - 1].ValorIss      := Leitor.rCampo(tcDe4, 'ValorIssqn');
-          NFSe.Servico.ItemServico.Items[I - 1].ValorTotal    := RoundTo(NFSe.Servico.ItemServico.Items[I - 1].Quantidade *
-                                                                 NFSe.Servico.ItemServico.Items[I - 1].ValorUnitario, - 2);
+          NFSe.Servico.ItemServico.Items[I - 1].ValorTotal    := RoundTo(Double(NFSe.Servico.ItemServico.Items[I - 1].Quantidade *
+                                                                 NFSe.Servico.ItemServico.Items[I - 1].ValorUnitario), - 2);
         end else
           Break;
       end;
